#43 Fix for various problems when starting up
After the transition to moving settings into the database, there were
numerous errors which occurred if a setting was not yet persisted. This
aims to fix that by being proactive about recognizing when a clean
install is done and registering all settings in addition to a new
method to persist a default setting if it is not yet registered.